Inline task join type normalization
