+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/tasks/$taskId/page.tsx
@@ -1,3 +1,4 @@
+import type { SelectUser } from "@superset/db/schema";
 import { ScrollArea } from "@superset/ui/scroll-area";
 import { Separator } from "@superset/ui/separator";
 import { eq, or } from "@tanstack/db";
@@ -7,11 +8,8 @@ import { createFileRoute, useNavigate } from "@tanstack/react-router";
 import { useMemo } from "react";
 import { apiTrpcClient } from "renderer/lib/api-trpc-client";
 import { useCollections } from "renderer/routes/_authenticated/providers/CollectionsProvider";
+import type { TaskWithStatus } from "../components/TasksView/hooks/useTasksTable";
 import { Route as TasksLayoutRoute } from "../layout";
-import {
-	normalizeTaskWithStatus,
-	type TaskWithStatus,
-} from "../utils/normalizeTaskWithStatus";
 import { ActivitySection } from "./components/ActivitySection";
 import { CommentInput } from "./components/CommentInput";
 import { EditableTitle } from "./components/EditableTitle";
@@ -67,7 +65,14 @@ function TaskDetailPage() {
 
 	const task: TaskWithStatus | null = useMemo(() => {
 		if (!taskData || taskData.length === 0) return null;
-		return normalizeTaskWithStatus(taskData[0]);
+		const task = taskData[0];
+		return {
+			...task,
+			assignee:
+				typeof task.assignee?.id === "string"
+					? (task.assignee as SelectUser)
+					: null,
+		};
 	}, [taskData]);
 	const taskFallbackQuery = useQuery({
 		queryKey: ["task-detail-fallback", taskId, isUuidTaskId ? "id" : "slug"],
