--- apps/desktop/src/renderer/routes/_authenticated/_dashboard/tasks/utils/normalizeTaskWithStatus/index.ts
@@ -1,4 +1,0 @@
-export {
-	normalizeTaskWithStatus,
-	type TaskWithStatus,
-} from "./normalizeTaskWithStatus";
