--- apps/desktop/src/renderer/routes/_authenticated/_dashboard/tasks/utils/normalizeTaskWithStatus/normalizeTaskWithStatus.ts
@@ -1,30 +1,0 @@
-import type {
-	SelectTask,
-	SelectTaskStatus,
-	SelectUser,
-} from "@superset/db/schema";
-
-type JoinedUser = Partial<SelectUser>;
-
-export type TaskWithStatus = SelectTask & {
-	status: SelectTaskStatus;
-	assignee: SelectUser | null;
-};
-
-type RawTaskWithStatus = SelectTask & {
-	status: SelectTaskStatus;
-	assignee: JoinedUser | null | undefined;
-};
-
-// Newer TanStack DB join inference widens left-joined records to partials.
-export function normalizeTaskWithStatus(
-	task: RawTaskWithStatus,
-): TaskWithStatus {
-	return {
-		...task,
-		assignee:
-			typeof task.assignee?.id === "string"
-				? (task.assignee as SelectUser)
-				: null,
-	};
-}
